Working injectPositive function - uses findIndex with a if negative predicate to check for the negative. If it's all positive, reduce the array to create a sum, copy the original via spread, push the sum to the newly created array, and return.
If any negative present, uses the map function to turn the negative (and all elements after it) into 0s in the new so that the array can be reduced to create a sum. After this, creates a clone of the original via spread, then splicing in the sum at the index after the negative's index.